Jogosultság kezelés, néhány új route felvétele, route-ok hibajavítása

+++ backend_P/routes/api.php
@@ -2,6 +2,15 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use \App\Http\Controllers\PeopleController;
+use \App\Http\Controllers\GenreController;
+use \App\Http\Controllers\MovieController;
+use \App\Http\Controllers\CommentController;
+use \App\Http\Controllers\RatingController;
+use \App\Http\Controllers\RoleController;
+use \App\Http\Controllers\UserController;
+use \App\Models\User;
+use \App\Models\Movie;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,45 +23,35 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+//************************************************************************************************************************************/
+/******* Genre/Tag *******************************************************************************************************************/
+/*************************************************************************************************************************************/
+
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-use \App\Http\Controllers\GenreController;
-//Route::resource('genre', GenreController::class);
-// Film műfajok lekérése
-// Műfajok lekérése
+
 //Route::get('genres', [GenreController::class, 'index']);
 Route::get('movies/tags',   [GenreController::class, 'index']);
 
 
 // Új műfaj létrehozása
-Route::post('genres', [GenreController::class, 'store'])->middleware('auth:sanctum');
+Route::post('genres', [GenreController::class, 'store'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Adott műfaj megtekintése
 Route::get('genres/{genre}', [GenreController::class, 'show']);
 
 // Adott műfaj szerkesztése
-Route::put('genres/{genre}', [GenreController::class, 'update'])->middleware('auth:sanctum');
+Route::put('genres/{genre}', [GenreController::class, 'update'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Adott műfaj törlése
-Route::delete('genres/{genre}', [GenreController::class, 'destroy'])->middleware('auth:sanctum');
-
-// Film műfajok lekérése
-//Route::get('movies/{movie}/genres', function (Movie $movie) {
-//    return $movie->genres;
-//});       EZ A BLOKK IDE VALÓ??
+Route::delete('genres/{genre}', [GenreController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 
-
-use \App\Http\Controllers\MovieController;
-use \App\Http\Controllers\CommentController;
-use \App\Http\Controllers\RatingController;
-use \App\Models\Movie;
-
-
-//Route::resource('movie', MovieController::class);
-//ezt lebontani 5 routra!
+/*************************************************************************************************************************************/
+/******* Movie ***********************************************************************************************************************/
+/*************************************************************************************************************************************/
 
 // Összes film lekérdezése
 Route::get('movies', [MovieController::class, 'index']);
@@ -62,114 +61,116 @@ Route::get('movies/{movie}/comments', [CommentController::class, 'indexByMovieID
 
 Route::get('movies/tag/{genre}', [MovieController::class, 'indexByGenre']);
 
+//Randon 4 film
+Route::get('movies/randommovies', [MovieController::class, 'randomMovies']);
 
 // Adott film megtekintése
 Route::get('movies/{movie}', [MovieController::class, 'show']);
 
-// Route::get('movies/{id}', function ($id) {
-//     $movie = Movie::findOrFail($id);
-
-//     $image = base64_encode(file_get_contents($movie->imageUrl));
-
-//     return response()->json([
-//         'id' => $movie->id,
-//         'name' => $movie->name,
-//         'release_year' => $movie->release_year,
-//         'description' => $movie->description,
-//         'imageUrl' => $movie->imageUrl,
-//         'length' => $movie->length,
-//         'origin' => $movie->origin
-//     ]);
-// });
 
 // Új film létrehozása
-Route::post('movies', [MovieController::class, 'store'])->middleware('auth:sanctum');
+Route::post('movies', [MovieController::class, 'store'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Adott film szerkesztése
-Route::put('movies/{movie}', [MovieController::class, 'update'])->middleware('auth:sanctum');
+Route::put('movies/{movie}', [MovieController::class, 'update'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Adott film törlése
-Route::delete('movies/{movie}', [MovieController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('movies/{movie}', [MovieController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
+
+// Filmes toplista
+Route::get('toplist', [MovieController::class, 'toplist']);
 
 
-Route::get('movies/toplist', [MovieController::class, 'toplist']);
+/*************************************************************************************************************************************/
+/******* Comment *********************************************************************************************************************/
+/*************************************************************************************************************************************/
 
-
-
-//Route::resource('movies.comments', CommentController::class);
-//Route::get('movie/{movie}/comment', [CommentController::class, 'index']);
-
+// Összes komment megtekintése
+Route::get('comments', [CommentController::class, 'index']);
 // Komment megtekintése
 Route::get('movies/{movie}/comments/{user}', [CommentController::class, 'show']);
 
 // Új komment hozzáadása
-Route::post('movies/comments', [CommentController::class, 'store'])->middleware('auth:sanctum');
+Route::post('movies/comments', [CommentController::class, 'store'])->middleware(['auth:sanctum', 'abilities:user']);
 
 // Komment szerkesztése
-Route::put('movies/comments/{comment}', [CommentController::class, 'update'])->middleware('auth:sanctum');
+Route::put('movies/comments/{comment}', [CommentController::class, 'update'])->middleware(['auth:sanctum', 'abilities:moderator']);
 
 // Komment törlése
-Route::delete('movies/comments/{comment}', [CommentController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('movies/comments/{comment}', [CommentController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:moderator']);
 
 
+/*************************************************************************************************************************************/
+/******* People **********************************************************************************************************************/
+/*************************************************************************************************************************************/
 
-use \App\Http\Controllers\PeopleController;
 // Színészek lekérése
 Route::get('people', [PeopleController::class, 'index']);
 
 // Új színész létrehozása
-Route::post('people', [PeopleController::class, 'store'])->middleware('auth:sanctum');
+Route::post('people', [PeopleController::class, 'store'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Színész megtekintése
 Route::get('people/{people}', [PeopleController::class, 'show']);
+// Színész megtekintése
+Route::get('people/name/{peoplename}', [PeopleController::class, 'showByName']);
 
 // Színész szerkesztése
-Route::put('people/{people}', [PeopleController::class, 'update'])->middleware('auth:sanctum');
+Route::put('people/{people}', [PeopleController::class, 'update'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Színész törlése
-Route::delete('people/{people}', [PeopleController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('people/{people}', [PeopleController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 
+/*************************************************************************************************************************************/
+/******* Rating **********************************************************************************************************************/
+/*************************************************************************************************************************************/
 
 // Új rating hozzáadása
-Route::post('movies/ratings', [RatingController::class, 'store'])->middleware('auth:sanctum');
-
-Route::get('comments', [CommentController::class, 'index']);
-
-use \App\Models\User;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\ValidationException;
-
-Route::post('/user/login', function (Request $request) {
-    $request->validate([
-        'email' => 'required|email',
-        'password' => 'required',
-    ]);
- 
-    $user = User::where('email', $request->email)->first();
- 
-    if (! $user || ! Hash::check($request->password, $user->password)) {
-        throw ValidationException::withMessages([
-            'email' => ['The provided credentials are incorrect.'],
-        ]);
-    }
- 
-    return $user->createToken("MovieCorner");
-});
+Route::post('movies/ratings', [RatingController::class, 'store'])->middleware(['auth:sanctum', 'abilities:user']);
 
 
-use \App\Http\Controllers\RoleController;
+/*************************************************************************************************************************************/
+/******* Role ************************************************************************************************************************/
+/*************************************************************************************************************************************/
+
 // Role-ok lekérdezése
-Route::get('role', [RoleController::class, 'index']);
+Route::get('roles', [RoleController::class, 'index']);
 
 // Új role létrehozása
-Route::post('role', [RoleController::class, 'store'])->middleware('auth:sanctum');
+Route::post('roles', [RoleController::class, 'store'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Role megtekintése
-Route::get('role/{role}', [RoleController::class, 'show']);
+Route::get('roles/{role}', [RoleController::class, 'show']);
 
 // Role szerkesztése
-Route::put('role/{role}', [RoleController::class, 'update'])->middleware('auth:sanctum');
+Route::put('roles/{role}', [RoleController::class, 'update'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
 
 // Role törlése
-Route::delete('role/{role}', [RoleController::class, 'destroy'])->middleware('auth:sanctum');
+Route::delete('roles/{role}', [RoleController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:admin']);
+
+
+/*************************************************************************************************************************************/
+/******* User ************************************************************************************************************************/
+/*************************************************************************************************************************************/
+
+// User login
+Route::post('/user/login', [UserController::class, 'login']);
+
+// User logout
+Route::post('/user/logout', [UserController::class, 'logout']);
+
+// User-ek lekérdezése
+Route::get('users', [UserController::class, 'index'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
+
+// Új user létrehozása
+Route::post('users', [UserController::class, 'store']);
+
+// Role megtekintése
+Route::get('users/{user}', [UserController::class, 'show'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
+
+// User szerkesztése
+Route::put('users/{user}', [UserController::class, 'update'])->middleware(['auth:sanctum', 'abilities:admin']);
+
+// User törlése
+Route::delete('users/{user}', [UserController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:admin,moderator']);
